myapp/serializers.py
from .models import Community, Post
from actstream.models import Action
from rest_framework import serializers
from rest_framework.serializers import ModelSerializer

class CommunitySerializer(serializers.ModelSerializer):
    class Meta:
        model = Community
        fields = '__all__'
        

class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = '__all__'
        
class ActionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Action
        fields = '__all__'
